backend/requirements.txt
Django==6.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-generativeai==0.3.0
python-decouple==3.8
requests==2.31.0
Pillow==10.1.0
